+++ app/src/main/java/io/hobaskos/event/eventapp/ui/profile/AttendingEventsFragment.java
@@ -4,16 +4,21 @@ import android.app.Fragment;
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.widget.DividerItemDecoration;
+import android.support.annotation.NonNull;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.hannesdorfmann.mosby.mvp.MvpFragment;
+
 import java.util.ArrayList;
 import java.util.List;
 
+import javax.inject.Inject;
+
+import io.hobaskos.event.eventapp.App;
 import io.hobaskos.event.eventapp.R;
 import io.hobaskos.event.eventapp.data.model.Event;
 import io.hobaskos.event.eventapp.ui.event.details.EventActivity;
@@ -23,20 +28,26 @@ import io.hobaskos.event.eventapp.ui.event.list.EventsAdapter;
  * Created by Magnus on 13.03.2017.
  */
 
-public class AttendingEventsFragment extends Fragment {
+public class AttendingEventsFragment extends MvpFragment<AttendingEventsView, AttendingEventsPresenter> implements AttendingEventsView{
+
     public static final String ARG_PAGE = "ARG_PAGE";
-    private static final String ARG_EVENT_LIST = "event-list";
-    private static final String ARG_EVENT_ID = "eventId";
+
+    private int mPage;
 
     private ArrayList<Event> eventAttending = new ArrayList<>();
 
+    @Inject
+    public AttendingEventsPresenter presenter;
 
-
-    public static AttendingEventsFragment newInstance(ArrayList<Event> event) {
+    public static AttendingEventsFragment newInstance(int page) {
         AttendingEventsFragment fragment = new AttendingEventsFragment();
+
         Bundle args = new Bundle();
-        args.putParcelableArrayList(ARG_EVENT_LIST, event);
+
+        args.putInt(ARG_PAGE, page);
+
         fragment.setArguments(args);
+
         return fragment;
     }
 
@@ -45,18 +56,32 @@ public class AttendingEventsFragment extends Fragment {
         super.onCreate(savedInstanceState);
 
         if (getArguments() != null) {
-            eventAttending = getArguments().getParcelableArrayList(ARG_EVENT_LIST);
+            mPage = getArguments().getInt(ARG_PAGE);
         }
+
+        //presenter.getEventUserAttending();
+
+
+    }
+
+    @NonNull
+    @Override
+    public AttendingEventsPresenter createPresenter() {
+        App.getInst().getComponent().inject(this);
+        return presenter;
     }
 
     @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
+    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
         View view = inflater.inflate(R.layout.fragment_page, container, false);
 
+
         Context context = view.getContext();
+
         RecyclerView recyclerView = (RecyclerView) view.findViewById(R.id.list);
+
         LinearLayoutManager linearLayoutManager = new LinearLayoutManager(context);
+
         recyclerView.setLayoutManager(linearLayoutManager);
         recyclerView.setAdapter(new EventsAdapter(eventAttending,context, event -> {
             Intent intent = new Intent(getActivity(), EventActivity.class);
@@ -69,5 +94,9 @@ public class AttendingEventsFragment extends Fragment {
     }
 
 
+    @Override
+    public void setEventAttending(List<Event> eventsAttending) {
+        eventAttending.addAll(eventsAttending);
+    }
 
 }//End of class
