+++ app/src/main/java/io/hobaskos/event/eventapp/ui/profile/AttendingEventsView.java
@@ -1,8 +1,15 @@
 package io.hobaskos.event.eventapp.ui.profile;
 
+import com.hannesdorfmann.mosby.mvp.MvpView;
+
+import java.util.List;
+
+import io.hobaskos.event.eventapp.data.model.Event;
+
 /**
  * Created by Magnus on 16.03.2017.
  */
 
-public class AttendingEventsView {
+public interface AttendingEventsView extends MvpView {
+    void setEventAttending(List<Event> eventAttending);
 }
